+++ decomp/ST.exe/functions/0065FE10/decomp.c
@@ -5,6 +5,7 @@ FUN_0065fe10(AiFltClassTy *param_1,undefined4 param_2,int param_3,short *param_4
 {
   short sVar1;
   undefined2 uVar2;
+  /* ST_PSEUDO[return_width_artifact]: candidate call-output artifact: verify return width, clobbers, or x87 state */
   undefined2 extraout_var;
   uint uVar3;
   DArrayTy *array;
@@ -17,7 +18,7 @@ FUN_0065fe10(AiFltClassTy *param_1,undefined4 param_2,int param_3,short *param_4
   DArrayTy *local_1c;
   DArrayTy *local_c;
   undefined4 local_8;
-  
+
   ppDVar5 = local_38;
   for (iVar4 = 0xb; iVar4 != 0; iVar4 = iVar4 + -1) {
     *ppDVar5 = (DArrayTy *)0x0;
@@ -27,6 +28,7 @@ FUN_0065fe10(AiFltClassTy *param_1,undefined4 param_2,int param_3,short *param_4
   local_8 = 0;
   param_1->field_00A7 = 0;
   uVar2 = AiFltClassTy::sub_0065D9C0(param_1);
+  /* ST_PSEUDO[return_width_artifact]: candidate call-output artifact: verify return width, clobbers, or x87 state */
   if (CONCAT22(extraout_var,uVar2) == 0) {
     return 0xffffffff;
   }
@@ -80,8 +82,9 @@ LAB_0065fec9:
       local_20 = 0;
       local_38[1] = array;
       local_1c = local_c;
+      /* ST_PSEUDO[packed_or_unaligned_piece]: expected named packed member, bit extract/compose, or unaligned load */
       if (((param_1->field_007D != -2) && (g_sTAllPlayers_007FA174 != (STAllPlayersC *)0x0)) &&
-         (this = thunk_FUN_0042b760((uint)(byte)param_1->field_0x24,
+         (this = thunk_FUN_0042b760(param_1->field_0x24,
                                     CONCAT22((short)((uint)local_c >> 0x10),param_1->field_007D)),
          this != (STGroupBoatC *)0x0)) {
         (*this->vtable->vfunc_08)(this,CASE_8,local_38);
