+++ decomp/ST.exe/functions/0075B800/decomp.c
@@ -9,9 +9,9 @@ uint FUN_0075b800(undefined4 *param_1,AnonShape_0075C180_8EED234B *param_2,int p
   int iVar1;
   uint uVar2;
   int *piVar3;
-  
+
   if (param_3 < param_5) {
-    iVar1 = FUN_0075b720(param_1,(uint)param_2,param_3,param_5);
+    iVar1 = FUN_0075b720(param_1,param_2,param_3,param_5);
     if (iVar1 == 0) {
       return 0xffffffff;
     }
@@ -24,7 +24,7 @@ uint FUN_0075b800(undefined4 *param_1,AnonShape_0075C180_8EED234B *param_2,int p
   if (*piVar3 < (int)uVar2) {
     do {
       if (iVar1 < 1) {
-        iVar1 = FUN_0075b720(param_1,(uint)param_2,iVar1,1);
+        iVar1 = FUN_0075b720(param_1,param_2,iVar1,1);
         if (iVar1 == 0) {
           return 0xffffffff;
         }
@@ -42,6 +42,7 @@ uint FUN_0075b800(undefined4 *param_1,AnonShape_0075C180_8EED234B *param_2,int p
   if (0x10 < param_5) {
     return 0;
   }
+  /* ST_PSEUDO[raw_pointer_offset]: candidate structure field after proof; otherwise retain buffer arithmetic */
   return (uint)*(byte *)((*(int *)(param_4 + 0x8c + param_5 * 4) - *(int *)(param_4 + param_5 * 4))
                          + *(int *)(param_4 + 0xd0) + 0x11 + uVar2);
 }
